refactor: User 계층 리팩토링
# 프로덕션 코드
- 예외 메세지 의존성 수정
## DAO
- fetch 메서드 : 2번 조회 -> 1번 조회
#테스트 코드
- gson -> objectMapper
- IntegrationTest -> UserIntegrationTest
- 예외 메세지 의존성 수정
