+++ src/main/java/com/flab/just_10_minutes/User/controller/UserController.java
@@ -17,7 +17,7 @@ public class UserController {
 
     private final UserService userService;
 
-    @PostMapping("/sign_up")
+    @PostMapping("/sign-up")
     public ResponseEntity<HttpStatus> signUp(@RequestBody @Valid UserDto userDto) {
         userService.save(UserDto.toDomain(userDto));
