+++ src/main/java/com/flab/just_10_minutes/User/service/UserService.java
@@ -6,7 +6,7 @@ import com.flab.just_10_minutes.Util.Exception.Business.BusinessException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import static com.flab.just_10_minutes.Util.contants.ResponseMessage.DUPLICATED_REGISTER;
+import static com.flab.just_10_minutes.Util.Exception.Business.BusinessException.DUPLICATED_REGISTER;
 
 @Service
 @RequiredArgsConstructor
@@ -23,17 +23,5 @@ public class UserService {
 
     public User getUserProfile(final String loginId) {
         return userDao.fetch(loginId);
-    }
-
-    public void update(final String loginId, final Long pointQuantity) {
-        if (!validateExistedUser(loginId)) {
-            throw new RuntimeException("not exist user");
-        }
-
-        int updateCount = userMapper.update(loginId, pointQuantity);
-
-        if (updateCount != 1) {
-            throw new RuntimeException("Fail update. Please retry");
-        }
     }
 }
