+++ src/test/java/com/flab/just_10_minutes/User/UserDtoTestFixture.java
@@ -2,7 +2,7 @@ package com.flab.just_10_minutes.User;
 
 import com.flab.just_10_minutes.User.dto.UserDto;
 
-class UserDtoTestFixture {
+public class UserDtoTestFixture {
 
     public static UserDto createTestUserDto(String loginId, String password, String phone, String address) {
         return UserDto.builder()
@@ -14,7 +14,7 @@ class UserDtoTestFixture {
     }
 
     public static final String NOT_EXIST_ID = "notExistId";
-    public static final String EXIST_ID = "testId";
+    public static final String EXIST_ID = "existId";
     public static final String PASSWORD = "testPassword";
     public static final String PHONE = "010-1234-5678";
     public static final String ADDRESS = "testAddress";
