+++ src/test/java/com/flab/just_10_minutes/User/UserTestFixture.java
@@ -12,7 +12,6 @@ public class UserTestFixture {
                 .phone("010-1234-5678")
                 .address("testAddress")
                 .role(User.ROLE.PUBLIC)
-                .point(0L)
                 .build();
     }
 
@@ -23,7 +22,6 @@ public class UserTestFixture {
                 .phone("010-1234-5678")
                 .address("testAddress")
                 .role(User.ROLE.PUBLIC)
-                .point(0L)
                 .build();
     }
 }
